SortedTable: add constructor using specified cols
Add constructor that uses the specified Columns for the resulting
SortedTable column definitions.  This is useful for providing column
definitions to a non-database-source dataset (e.g. XML file).

The other constructors use the column definitions from the
table's metadata.  A non-database-source dataset's column definitions
are "unknown", so data comparisons and sorting results are not always as
desired.