B427: Fix normalization of schemas in DatabaseDataSet
Schemas should be stored in upper case only when case sensitivity is
disabled. Otherwise case information required for case sensitivity is
lost.

Add case sensitivity test for DatabaseDataSet.
This verifies that schemas are handled correctly when case sensitivity
is disabled. It does so by ensuring that 'getTables' method of the
metadata handler is called for only one schema. This works because
DatabaseDataSet caches tables. The cached metadata should be re-used
when requesting metadata for a lower-case qualified table after having
done the same for the table in upper-case.