B428: Fix value casting in IsActualWithinToleranceOfExpectedTimestampValueComparer
Actual and expected values are now cast to the expected type before
converting to milliseconds for the comparison. This allows
non-Timestamp objects to used with
IsActualWithinToleranceOfExpectedTimestampValueComparer. This is
useful when comparing data of a known type to data with unknown types.

Add testStringExpectedTimestampActual to
IsActualWithinToleranceOfExpectedTimestampValueComparerTest.
This test emulates the situation where untyped expected data is being
compared to typed actual data.